Refactor file structure in ui package
