+++ libraries/yas-ui/src/form/CheckboxGroupField.tsx
@@ -1,10 +1,10 @@
 import { useId, type ReactNode } from "react";
-import type { FieldProps } from "./rcf";
+import type { FieldProps } from "./shared/rcf";
 import {
   FormControl,
   FormControlErrors,
   FormControlLabel,
-} from "./FormControl";
+} from "./shared/FormControl";
 import { CheckboxField } from "./CheckboxField";
 
 export interface CheckboxGroupOption<Value> {
