+++ libraries/yas-ui/src/form/MultiSelectField.tsx
@@ -12,12 +12,12 @@ import {
   SelectScrollDownButton,
 } from "../atoms/Select";
 import { SelectRoot } from "../atoms/Select";
-import type { FieldProps } from "./rcf";
+import type { FieldProps } from "./shared/rcf";
 import {
   FormControl,
   FormControlLabel,
   FormControlErrors,
-} from "./FormControl";
+} from "./shared/FormControl";
 
 export interface MultiSelectOption<Value> {
   value: Value;
