+++ libraries/yas-ui/src/form/SingleSelectField.tsx
@@ -13,12 +13,12 @@ import {
   SelectValue,
   SelectTriggerIcon,
 } from "../atoms/Select";
-import type { FieldProps } from "./rcf";
+import type { FieldProps } from "./shared/rcf";
 import {
   FormControl,
   FormControlLabel,
   FormControlErrors,
-} from "./FormControl";
+} from "./shared/FormControl";
 
 export type SingleSelectOption<Value> = {
   value: Value;
