+++ libraries/yas-ui/src/form/TextField.tsx
@@ -6,13 +6,13 @@ import { Cross1Icon } from "@yas/icons";
 import { InputArea, InputRoot, InputSlot } from "../atoms/Input";
 import { Button } from "../atoms/Button";
 import { CircularProgress } from "../atoms/CircularProgress";
-import type { FieldProps } from "./rcf";
+import type { FieldProps } from "./shared/rcf";
 import {
   FormControl,
   FormControlErrors,
   FormControlLabel,
   type FormControlProps,
-} from "./FormControl";
+} from "./shared/FormControl";
 import * as styles from "./TextField.css";
 
 export interface TextFieldProps
