+++ libraries/yas-ui/src/form/composed/BaseForm.ts
@@ -1,8 +1,8 @@
 import { z } from "@yas/validate";
-import { TextField } from "./TextField";
-import { NumberField } from "./NumberField";
-import { createForm } from "./rcf";
-import { BasicFormLayout } from "./BasicFormLayout";
+import { TextField } from "../TextField";
+import { NumberField } from "../NumberField";
+import { BasicFormLayout } from "../BasicFormLayout";
+import { createForm } from "../shared/rcf";
 
 export const BaseForm = createForm((options) =>
   options
