+++ libraries/yas-ui/src/index.ts
@@ -19,11 +19,11 @@ export * from "./layout/Bounds";
 export * from "./layout/Stack";
 export * from "./layout/Container";
 export * from "./layout/Void";
-export * from "./form/BaseForm";
+export * from "./form/shared/rcf";
+export * from "./form/composed/BaseForm";
+export * from "./form/shared/FormControl";
 export * from "./form/BasicFormLayout";
-export * from "./form/FormControl";
 export * from "./form/Pagination";
-export * from "./form/rcf";
 export * from "./form/CheckboxField";
 export * from "./form/CheckboxGroupField";
 export * from "./form/DatePicker";
